requirements.txt
apache-airflow==2.9.1
pandas
requests
psycopg2-binary
python-dotenv
